Just C++

+++ TriangleClass-2.cpp
@@ -2,9 +2,9 @@
 #include<math.h>
 using namespace std;
 struct Triangle {
-	double x1,y1;
-	double x2,y2;
-	double x3,y3;
+	long double x1,y1;
+	long double x2,y2;
+	long double x3,y3;
 	void input() {
 		cin >> x1 >> y1;
 		cin >> x2 >> y2;
@@ -12,7 +12,7 @@ struct Triangle {
 	}
 };
 
-double check(Triangle &tg,double a,double b,double c) {
+double check(Triangle &tg,long double a,long double b,long double c) {
 	if(a+b<=c || a+c<=b || b+c<=a) {
 		return 0;
 	}	else {
@@ -27,13 +27,13 @@ int main() {
 	while(t--) {
 		Triangle tg;
 		tg.input();
-		double a=sqrt((tg.x1-tg.x2)*(tg.x1-tg.x2)+(tg.y1-tg.y2)*(tg.y1-tg.y2));
-		double b=sqrt((tg.x1-tg.x3)*(tg.x1-tg.x3)+(tg.y1-tg.y3)*(tg.y1-tg.y3));
-		double c=sqrt((tg.x2-tg.x3)*(tg.x2-tg.x3)+(tg.y2-tg.y3)*(tg.y2-tg.y3));
+		long double a=sqrt((tg.x1-tg.x2)*(tg.x1-tg.x2)+(tg.y1-tg.y2)*(tg.y1-tg.y2));
+		long double b=sqrt((tg.x1-tg.x3)*(tg.x1-tg.x3)+(tg.y1-tg.y3)*(tg.y1-tg.y3));
+		long double c=sqrt((tg.x2-tg.x3)*(tg.x2-tg.x3)+(tg.y2-tg.y3)*(tg.y2-tg.y3));
 		if(check(tg,a,b,c)==0) {
 			cout << "INVALID" << endl;
 		}	else {
-			double kq = check(tg,a,b,c);
+			long double kq = check(tg,a,b,c);
 			cout << setprecision(2) << fixed << kq << endl;
 		}
 	}
